Complete WishList separate component
